dense: expose SceneDensify data structures and functions (#476)
