+++ libs/MVS.h
@@ -41,6 +41,7 @@
 #include "Math/Common.h"
 #include "MVS/Common.h"
 #include "MVS/Scene.h"
+#include "MVS/SceneDensify.h"
 
 
 #endif // _MVS_MVS_H_
