+++ libs/MVS/SceneDensify.cpp
@@ -31,6 +31,7 @@
 
 #include "Common.h"
 #include "Scene.h"
+#include "SceneDensify.h"
 // MRF: view selection
 #include "../Math/TRWS/MRFEnergy.h"
 // CGAL: depth-map initialization
@@ -125,45 +126,7 @@ inline float Conf2Weight(float conf, Depth depth) {
 /*----------------------------------------------------------------*/
 
 
-
 // S T R U C T S ///////////////////////////////////////////////////
-
-// structure used to compute all depth-maps
-class DepthMapsData
-{
-public:
-	DepthMapsData(Scene& _scene);
-	~DepthMapsData();
-
-	bool SelectViews(IIndexArr& images, IIndexArr& imagesMap, IIndexArr& neighborsMap);
-	bool SelectViews(DepthData& depthData);
-	bool InitViews(DepthData& depthData, IIndex idxNeighbor, IIndex numNeighbors);
-	bool InitDepthMap(DepthData& depthData);
-	bool EstimateDepthMap(IIndex idxImage);
-
-	bool RemoveSmallSegments(DepthData& depthData);
-	bool GapInterpolation(DepthData& depthData);
-
-	bool FilterDepthMap(DepthData& depthData, const IIndexArr& idxNeighbors, bool bAdjust=true);
-	void FuseDepthMaps(PointCloud& pointcloud, bool bEstimateColor, bool bEstimateNormal);
-
-protected:
-	static void* STCALL ScoreDepthMapTmp(void*);
-	static void* STCALL EstimateDepthMapTmp(void*);
-	static void* STCALL EndDepthMapTmp(void*);
-
-public:
-	Scene& scene;
-
-	DepthDataArr arrDepthData;
-
-	// used internally to estimate the depth-maps
-	Image8U::Size prevDepthMapSize; // remember the size of the last estimated depth-map
-	Image8U::Size prevDepthMapSizeTrg; // ... same for target image
-	DepthEstimator::MapRefArr coords; // map pixel index to zigzag matrix coordinates
-	DepthEstimator::MapRefArr coordsTrg; // ... same for target image
-};
-/*----------------------------------------------------------------*/
 
 
 DepthMapsData::DepthMapsData(Scene& _scene)
@@ -1560,32 +1523,54 @@ void DepthMapsData::FuseDepthMaps(PointCloud& pointcloud, bool bEstimateColor, b
 
 // S T R U C T S ///////////////////////////////////////////////////
 
-struct DenseDepthMapData {
-	Scene& scene;
-	IIndexArr images;
-	IIndexArr neighborsMap;
-	DepthMapsData detphMaps;
-	volatile Thread::safe_t idxImage;
-	SEACAVE::EventQueue events; // internal events queue (processed by the working threads)
-	Semaphore sem;
-	CAutoPtr<Util::Progress> progress;
-
-	DenseDepthMapData(Scene& _scene)
-		: scene(_scene), detphMaps(_scene), idxImage(0), sem(1) {}
-
-	void SignalCompleteDepthmapFilter() {
-		ASSERT(idxImage > 0);
-		if (Thread::safeDec(idxImage) == 0)
-			sem.Signal((unsigned)images.GetSize()*2);
-	}
-};
-
 static void* DenseReconstructionEstimateTmp(void*);
 static void* DenseReconstructionFilterTmp(void*);
-bool Scene::DenseReconstruction()
-{
-	DenseDepthMapData data(*this);
 
+bool Scene::DenseReconstruction() {
+	DenseDepthMapData data(*this);
+	
+	// estimate depth-maps
+	if (!ComputeDepthMaps(data))
+		return false;
+	
+	// fuse all depth-maps
+	pointcloud.Release();
+	data.depthMaps.FuseDepthMaps(pointcloud, OPTDENSE::nEstimateColors == 2, OPTDENSE::nEstimateNormals == 2);
+	#if TD_VERBOSE != TD_VERBOSE_OFF
+	if (g_nVerbosityLevel > 2) {
+		// print number of points with 3+ views
+		size_t nPoints1m(0), nPoints2(0), nPoints3p(0);
+		FOREACHPTR(pViews, pointcloud.pointViews) {
+			switch (pViews->GetSize())
+			{
+			case 0:
+			case 1:
+				++nPoints1m;
+				break;
+			case 2:
+				++nPoints2;
+				break;
+			default:
+				++nPoints3p;
+			}
+		}
+		VERBOSE("Dense point-cloud composed of:\n\t%u points with 1- views\n\t%u points with 2 views\n\t%u points with 3+ views", nPoints1m, nPoints2, nPoints3p);
+	}
+	#endif
+
+	if (!pointcloud.IsEmpty()) {
+		if (pointcloud.colors.IsEmpty() && OPTDENSE::nEstimateColors == 1)
+			EstimatePointColors(images, pointcloud);
+		if (pointcloud.normals.IsEmpty() && OPTDENSE::nEstimateNormals == 1)
+			EstimatePointNormals(images, pointcloud);
+	}
+	return true;
+} // DenseReconstruction
+/*----------------------------------------------------------------*/
+
+// do first half of dense reconstruction: depth map computation
+// results are saved to "data"
+bool Scene::ComputeDepthMaps(DenseDepthMapData& data) {
 	{
 	// maps global view indices to our list of views to be processed
 	IIndexArr imagesMap;
@@ -1665,8 +1650,8 @@ bool Scene::DenseReconstruction()
 		#endif
 			const IIndex idxImage(data.images[idx]);
 			ASSERT(imagesMap[idxImage] != NO_ID);
-			DepthData& depthData(data.detphMaps.arrDepthData[idxImage]);
-			if (!data.detphMaps.SelectViews(depthData)) {
+			DepthData& depthData(data.depthMaps.arrDepthData[idxImage]);
+			if (!data.depthMaps.SelectViews(depthData)) {
 				#ifdef DENSE_USE_OPENMP
 				#pragma omp critical
 				#endif
@@ -1680,7 +1665,7 @@ bool Scene::DenseReconstruction()
 			data.images.RemoveAt(idx);
 		}
 		// globally select a target view for each reference image
-		if (OPTDENSE::nNumViews == 1 && !data.detphMaps.SelectViews(data.images, imagesMap, data.neighborsMap)) {
+		if (OPTDENSE::nNumViews == 1 && !data.depthMaps.SelectViews(data.images, imagesMap, data.neighborsMap)) {
 			VERBOSE("error: no valid images to be dense reconstructed");
 			return false;
 		}
@@ -1738,40 +1723,8 @@ bool Scene::DenseReconstruction()
 			return false;
 		data.progress.Release();
 	}
-
-	// fuse all depth-maps
-	pointcloud.Release();
-	data.detphMaps.FuseDepthMaps(pointcloud, OPTDENSE::nEstimateColors == 2, OPTDENSE::nEstimateNormals == 2);
-	#if TD_VERBOSE != TD_VERBOSE_OFF
-	if (g_nVerbosityLevel > 2) {
-		// print number of points with 3+ views
-		size_t nPoints1m(0), nPoints2(0), nPoints3p(0);
-		FOREACHPTR(pViews, pointcloud.pointViews) {
-			switch (pViews->GetSize())
-			{
-			case 0:
-			case 1:
-				++nPoints1m;
-				break;
-			case 2:
-				++nPoints2;
-				break;
-			default:
-				++nPoints3p;
-			}
-		}
-		VERBOSE("Dense point-cloud composed of:\n\t%u points with 1- views\n\t%u points with 2 views\n\t%u points with 3+ views", nPoints1m, nPoints2, nPoints3p);
-	}
-	#endif
-
-	if (!pointcloud.IsEmpty()) {
-		if (pointcloud.colors.IsEmpty() && OPTDENSE::nEstimateColors == 1)
-			EstimatePointColors(images, pointcloud);
-		if (pointcloud.normals.IsEmpty() && OPTDENSE::nEstimateNormals == 1)
-			EstimatePointNormals(images, pointcloud);
-	}
 	return true;
-} // DenseReconstruction
+} // ComputeDepthMaps
 /*----------------------------------------------------------------*/
 
 void* DenseReconstructionEstimateTmp(void* arg) {
@@ -1781,8 +1734,7 @@ void* DenseReconstructionEstimateTmp(void* arg) {
 }
 
 // initialize the dense reconstruction with the sparse point cloud
-void Scene::DenseReconstructionEstimate(void* pData)
-{
+void Scene::DenseReconstructionEstimate(void* pData) {
 	DenseDepthMapData& data = *((DenseDepthMapData*)pData);
 	while (true) {
 		CAutoPtr<Event> evt(data.events.GetEvent());
@@ -1798,10 +1750,10 @@ void Scene::DenseReconstructionEstimate(void* pData)
 			}
 			// select views to reconstruct the depth-map for this image
 			const IIndex idx = data.images[evtImage.idxImage];
-			DepthData& depthData(data.detphMaps.arrDepthData[idx]);
+			DepthData& depthData(data.depthMaps.arrDepthData[idx]);
 			// init images pair: reference image and the best neighbor view
 			ASSERT(data.neighborsMap.IsEmpty() || data.neighborsMap[evtImage.idxImage] != NO_ID);
-			if (!data.detphMaps.InitViews(depthData, data.neighborsMap.IsEmpty()?NO_ID:data.neighborsMap[evtImage.idxImage], OPTDENSE::nNumViews)) {
+			if (!data.depthMaps.InitViews(depthData, data.neighborsMap.IsEmpty()?NO_ID:data.neighborsMap[evtImage.idxImage], OPTDENSE::nNumViews)) {
 				// process next image
 				data.events.AddEvent(new EVTProcessImage((IIndex)Thread::safeInc(data.idxImage)));
 				break;
@@ -1830,7 +1782,7 @@ void Scene::DenseReconstructionEstimate(void* pData)
 			data.events.AddEvent(new EVTProcessImage((uint32_t)Thread::safeInc(data.idxImage)));
 			// extract depth map
 			data.sem.Wait();
-			data.detphMaps.EstimateDepthMap(data.images[evtImage.idxImage]);
+			data.depthMaps.EstimateDepthMap(data.images[evtImage.idxImage]);
 			data.sem.Signal();
 			if (OPTDENSE::nOptimize & OPTDENSE::OPTIMIZE) {
 				// optimize depth-map
@@ -1844,7 +1796,7 @@ void Scene::DenseReconstructionEstimate(void* pData)
 		case EVT_OPTIMIZEDEPTHMAP: {
 			const EVTOptimizeDepthMap& evtImage = *((EVTOptimizeDepthMap*)(Event*)evt);
 			const IIndex idx = data.images[evtImage.idxImage];
-			DepthData& depthData(data.detphMaps.arrDepthData[idx]);
+			DepthData& depthData(data.depthMaps.arrDepthData[idx]);
 			#if TD_VERBOSE != TD_VERBOSE_OFF
 			// save depth map as image
 			if (g_nVerbosityLevel > 3)
@@ -1853,13 +1805,13 @@ void Scene::DenseReconstructionEstimate(void* pData)
 			// apply filters
 			if (OPTDENSE::nOptimize & (OPTDENSE::REMOVE_SPECKLES)) {
 				TD_TIMER_START();
-				if (data.detphMaps.RemoveSmallSegments(depthData)) {
+				if (data.depthMaps.RemoveSmallSegments(depthData)) {
 					DEBUG_ULTIMATE("Depth-map %3u filtered: remove small segments (%s)", idx, TD_TIMER_GET_FMT().c_str());
 				}
 			}
 			if (OPTDENSE::nOptimize & (OPTDENSE::FILL_GAPS)) {
 				TD_TIMER_START();
-				if (data.detphMaps.GapInterpolation(depthData)) {
+				if (data.depthMaps.GapInterpolation(depthData)) {
 					DEBUG_ULTIMATE("Depth-map %3u filtered: gap interpolation (%s)", idx, TD_TIMER_GET_FMT().c_str());
 				}
 			}
@@ -1870,7 +1822,7 @@ void Scene::DenseReconstructionEstimate(void* pData)
 		case EVT_SAVEDEPTHMAP: {
 			const EVTSaveDepthMap& evtImage = *((EVTSaveDepthMap*)(Event*)evt);
 			const IIndex idx = data.images[evtImage.idxImage];
-			DepthData& depthData(data.detphMaps.arrDepthData[idx]);
+			DepthData& depthData(data.depthMaps.arrDepthData[idx]);
 			#if TD_VERBOSE != TD_VERBOSE_OFF
 			// save depth map as image
 			if (g_nVerbosityLevel > 2) {
@@ -1916,7 +1868,7 @@ void Scene::DenseReconstructionFilter(void* pData)
 		case EVT_FILTERDEPTHMAP: {
 			const EVTFilterDepthMap& evtImage = *((EVTFilterDepthMap*)(Event*)evt);
 			const IIndex idx = data.images[evtImage.idxImage];
-			DepthData& depthData(data.detphMaps.arrDepthData[idx]);
+			DepthData& depthData(data.depthMaps.arrDepthData[idx]);
 			if (!depthData.IsValid()) {
 				data.SignalCompleteDepthmapFilter();
 				break;
@@ -1927,7 +1879,7 @@ void Scene::DenseReconstructionFilter(void* pData)
 			IIndexArr idxNeighbors(0, depthData.neighbors.GetSize());
 			FOREACH(n, depthData.neighbors) {
 				const IIndex idxView = depthData.neighbors[n].idx.ID;
-				DepthData& depthDataPair = data.detphMaps.arrDepthData[idxView];
+				DepthData& depthDataPair = data.depthMaps.arrDepthData[idxView];
 				if (!depthDataPair.IsValid())
 					continue;
 				if (depthDataPair.IncRef(ComposeDepthFilePath(idxView, "dmap")) == 0) {
@@ -1940,14 +1892,14 @@ void Scene::DenseReconstructionFilter(void* pData)
 					break;
 			}
 			// filter the depth-map for this image
-			if (data.detphMaps.FilterDepthMap(depthData, idxNeighbors, OPTDENSE::bFilterAdjust)) {
+			if (data.depthMaps.FilterDepthMap(depthData, idxNeighbors, OPTDENSE::bFilterAdjust)) {
 				// load the filtered maps after all depth-maps were filtered
 				data.events.AddEvent(new EVTAdjustDepthMap(evtImage.idxImage));
 			}
 			// unload referenced depth-maps
 			FOREACHPTR(pIdxNeighbor, idxNeighbors) {
 				const IIndex idxView = depthData.neighbors[*pIdxNeighbor].idx.ID;
-				DepthData& depthDataPair = data.detphMaps.arrDepthData[idxView];
+				DepthData& depthDataPair = data.depthMaps.arrDepthData[idxView];
 				depthDataPair.DecRef();
 			}
 			depthData.DecRef();
@@ -1957,7 +1909,7 @@ void Scene::DenseReconstructionFilter(void* pData)
 		case EVT_ADJUSTDEPTHMAP: {
 			const EVTAdjustDepthMap& evtImage = *((EVTAdjustDepthMap*)(Event*)evt);
 			const IIndex idx = data.images[evtImage.idxImage];
-			DepthData& depthData(data.detphMaps.arrDepthData[idx]);
+			DepthData& depthData(data.depthMaps.arrDepthData[idx]);
 			ASSERT(depthData.IsValid());
 			data.sem.Wait();
 			// load filtered maps
